AutoSTP/backend/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart
gunicorn

# Database
sqlalchemy
asyncpg
alembic

# Authentication and security
python-jose[cryptography]
bcrypt
python-dotenv

# Pydantic and validation
pydantic
pydantic-settings
email-validator

# OpenRouter API (LLM integration)
openai

# Document processing
python-docx
PyPDF2
pdfplumber
reportlab

# Rate limiting
slowapi

# Templating
jinja2

# Testing
pytest
pytest-asyncio
pytest-mock
